[CodeEditor] fix for xbuild and setup and deploy the debugger via msbuild targets for OSX and Windows
Due to differences in the way xbuild and msbuild handle empty projects (xbuild invokes
the compiler with no source files and the build fails), add a little prebuild step
that touches the target assembly so that the build step is skipped (it doesn't have
to run anyway).

After the build, take the referenced assemblies and put them in the right place
for the debugger.

Add a "BuildDebuggerProject" target that builds and deploys the unity debugger
project (called with msbuild/xbuild /t:BuildDebuggerProject