[debugger] debugger client/server and sources listing implementation
Implements a udp client/server for communication between
the debugger app and the Unity editor, to query information
directly without having to rely on sdb being attached.
This is based on a rewrite of the completion service
client/server udp implementation.
Implements sources listing, for inspection and breakpoint
setting before the debugger is attached.